Use only headers with allocation Google tests
Delete old headers and tests.
Update build targets.
Rename tests and headers (possibly contentious)
