+++ WORKSPACE.bazel
@@ -12,18 +12,6 @@ http_archive(
 )
 
 http_archive(
-  name = "googletest",
-  urls = ["https://github.com/google/googletest/archive/5ab508a01f9eb089207ee87fd547d290da39d015.zip"],
-  strip_prefix = "googletest-5ab508a01f9eb089207ee87fd547d290da39d015",
-)
-
-http_archive(
-    name = "catch2",
-    strip_prefix = "Catch2-3.3.2",
-    urls = ["https://github.com/catchorg/Catch2/archive/refs/tags/v3.3.2.tar.gz"],
-)
-
-http_archive(
   name = "com_google_googletest",
   urls = ["https://github.com/google/googletest/archive/5ab508a01f9eb089207ee87fd547d290da39d015.zip"],
   strip_prefix = "googletest-5ab508a01f9eb089207ee87fd547d290da39d015",
